pstmgmt/requirements.txt
Django==3.1.3
fuzzywuzzy==0.18.0
opencv_python==4.4.0.46
opencv-contrib-python==4.4.0.46
djangorestframework==3.12.2
Pillow==8.0.1
python-Levenshtein==0.12.2
twilio==6.49.0
